requirements.txt
# Multi-Agent Ecosystem Requirements
# ====================================
# PhD-Level Production Dependencies
# Complete integration with existing CESAR infrastructure

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

# PostgreSQL Vector Extension
pgvector==0.2.4

# Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# LLM Integrations
openai==1.7.2
google-generativeai==0.3.2
anthropic==0.8.1
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2

# Vector Search & Embeddings
faiss-cpu==1.7.4
numpy==1.24.3
scipy==1.11.4

# Data Processing & Analysis
pandas==2.1.4
numpy==1.24.3

# Financial Data (FinPsyMCP)
yfinance==0.2.35
pandas-datareader==0.10.0
fredapi==0.5.1
alpha-vantage==2.3.1
ccxt==4.2.8  # Cryptocurrency exchange data
textblob==0.17.1  # Sentiment analysis
prophet==1.1.5  # Time-series forecasting
scikit-learn==1.3.2  # ML algorithms

# Workflow Orchestration (Prefect)
prefect==2.14.9
prefect-sqlalchemy==0.4.0

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# HTTP Client
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Rate Limiting
slowapi==0.1.9

# Web Scraping (ScraperAgents)
beautifulsoup4==4.12.2
lxml==5.1.0
selenium==4.16.0

# Document Processing (LexMCP)
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Data Validation
pydantic==2.5.3
email-validator==2.1.0

# Configuration Management
python-dotenv==1.0.0

# Logging & Tracing
structlog==24.1.0
python-json-logger==2.0.7

# Task Queue & Background Jobs
celery==5.3.4
kombu==5.3.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
jinja2==3.1.3

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0  # For API testing

# Code Quality (Development)
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3

# Documentation (Development)
mkdocs==1.5.3
mkdocs-material==9.5.3

# CORS Middleware
python-cors==1.0.0

# WebSocket Support
websockets==12.0

# Compression
brotli==1.1.0

# JSON Handling
orjson==3.9.12

# UUID Generation
uuid==1.30

# Type Checking
typing-extensions==4.9.0

# Asyncio Utilities
asyncio==3.4.3
aiofiles==23.2.1

# Cryptography (GambinoSecurityMCP)
cryptography==42.0.0
pycryptodome==3.19.0

# Patent Search (InnoMCP)
patentsview==0.2.3

# Music Composition (CreativeMCP)
music21==9.1.0  # Music theory library

# Image Processing (CreativeMCP)
pillow==10.2.0

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2

# Workflow Parsing (PydiniRedEnterprise)
pyyaml==6.0.1
jsonschema==4.20.0

# Circuit Breaker Pattern
pybreaker==1.0.1

# Retry Logic
tenacity==8.2.3

# Progress Bars
tqdm==4.66.1

# CLI
click==8.1.7
rich==13.7.0

# Health Checks
healthcheck==1.3.3

# Timezone Handling
tzdata==2023.4

# URL Parsing
urllib3==2.1.0

# Markdown Processing
markdown==3.5.2

# Code Formatting
autopep8==2.0.4

# Static Type Checking
types-requests==2.31.0.20240106
types-redis==4.6.0.20240106

# Performance Profiling
py-spy==0.3.14
memory-profiler==0.61.0

# Excel/CSV Processing
openpyxl==3.1.2
xlsxwriter==3.1.9

# Calendar & Date Utilities
holidays==0.38

# IP Address Handling
ipaddress==1.0.23

# Regular Expression Utilities
regex==2023.12.25

# Caching
cachetools==5.3.2

# Concurrent Execution
concurrent-log-handler==0.9.24

# Environment Management
environs==10.3.0

# Secrets Management
keyring==24.3.0

# Resource Limits
resource==0.2.1

# Signal Handling
signal==0.1.0

# Graceful Shutdown
signal-graceful-shutdown==0.1.0
